Refactor: extract ApiKeyEditDialog
